LSTM: new command line option: -lstmonly. Rewritten help screen.

+++ lstm/LstmModelContainer.hpp
@@ -39,7 +39,7 @@ public:
     size_t horizon);
 
   void next();
-  int getp();
+  float getp();
   void mix(Mixer& m);
   void update() override;
